requirements.txt
colorama==0.4.6
python-dotenv==1.0.1
pytube==15.0.0
tqdm==4.66.5
youtube-dl==2021.12.17
